project2_sentiment/requirements.txt
transformers==4.41.1
torch==2.3.0
yfinance==0.2.40
requests==2.32.2
pandas==2.2.2
numpy==1.26.4
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
scikit-learn==1.5.0
